users/views.py
from rest_framework import validators, viewsets
from rest_framework import permissions
from .serializes import UserSerializer
from .models import User

class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer
    permission_classes = [permissions.AllowAny]